Feature 43 contract crud

See merge request !15

+++ src/main/java/ua/com/foxminded/serviceacc/model/Money.java
@@ -15,60 +15,60 @@ import org.hibernate.annotations.Parameter;
 @Table(name = "money")
 public class Money {
 
-	@Id
-	@GenericGenerator(name = "generator", strategy = "org.hibernate.id.enhanced.SequenceStyleGenerator", parameters = {
-			@Parameter(name = "sequence_name", value = "money_id_seq"), @Parameter(name = "initial_value", value = "1"),
-			@Parameter(name = "increment_size", value = "1") })
-	@GeneratedValue(generator = "generator")
-	@Column(name = "id", unique = true, nullable = false)
-	private Long id;
+    @Id
+    @GenericGenerator(name = "generator", strategy = "org.hibernate.id.enhanced.SequenceStyleGenerator", parameters = {
+            @Parameter(name = "sequence_name", value = "money_id_seq"), @Parameter(name = "initial_value", value = "1"),
+            @Parameter(name = "increment_size", value = "1") })
+    @GeneratedValue(generator = "generator")
+    @Column(name = "id", unique = true, nullable = false)
+    private Long id;
 
-	@Column(name = "currency")
-	private Currency currency;
+    @Column(name = "currency")
+    private Currency currency;
 
-	@Column(name = "amount")
-	private int amount;
+    @Column(name = "amount")
+    private Long amount;
 
-	@Column(name = "active", nullable = false)
-	private boolean active = true;
+    @Column(name = "active", nullable = false)
+    private boolean active = true;
 
-	public Money() {
-	}
+    public Money() {
+    }
 
-	public Money(Currency currency, int amount) {
-		this.currency = currency;
-		this.amount = amount;
-	}
+    public Money(Currency currency, Long amount) {
+        this.currency = currency;
+        this.amount = amount;
+    }
 
-	public long getId() {
-		return id;
-	}
+    public long getId() {
+        return id;
+    }
 
-	public void setId(long id) {
-		this.id = id;
-	}
+    public void setId(long id) {
+        this.id = id;
+    }
 
-	public Currency getCurrency() {
-		return currency;
-	}
+    public Currency getCurrency() {
+        return currency;
+    }
 
-	public void setCurrency(Currency currency) {
-		this.currency = currency;
-	}
+    public void setCurrency(Currency currency) {
+        this.currency = currency;
+    }
 
-	public int getAmount() {
-		return amount;
-	}
+    public Long getAmount() {
+        return amount;
+    }
 
-	public void setAmount(int amount) {
-		this.amount = amount;
-	}
+    public void setAmount(Long amount) {
+        this.amount = amount;
+    }
 
-	public boolean isActive() {
-		return active;
-	}
+    public boolean isActive() {
+        return active;
+    }
 
-	public void setActive(boolean active) {
-		this.active = active;
-	}
+    public void setActive(boolean active) {
+        this.active = active;
+    }
 }
